tools/release/requirements.txt
PyGithub
pytest
pytest-mock
